requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==6.0.3
django-allauth==65.14.3
django-crispy-forms==2.6
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
gunicorn
whitenoise
dj-database-url
psycopg2-binary
xhtml2pdf
